refactor: 인수 테스트에서 카드 & 문제집 생성 메서드 AuthAcceptanceTest로 이동
