+++ backend/src/test/java/botobo/core/quiz/QuizAcceptanceTest.java
@@ -166,38 +166,4 @@ public class QuizAcceptanceTest extends AuthAcceptanceTest {
         assertThat(secondResponse.statusCode()).isEqualTo(HttpStatus.OK.value());
         assertThat(secondQuizResponses.size()).isEqualTo(10);
     }
-
-    public ExtractableResponse<Response> 문제집_생성_요청(AdminWorkbookRequest adminWorkbookRequest) {
-        return RestAssured.given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .auth().oauth2(로그인되어_있음().getAccessToken())
-                .body(adminWorkbookRequest)
-                .when().post("/api/admin/workbooks")
-                .then().log().all()
-                .extract();
-    }
-
-    public void 여러개_문제집_생성_요청(List<AdminWorkbookRequest> requests) {
-        for (AdminWorkbookRequest request : requests) {
-            RestAssured.given().log().all()
-                    .contentType(MediaType.APPLICATION_JSON_VALUE)
-                    .auth().oauth2(로그인되어_있음().getAccessToken())
-                    .body(request)
-                    .when().post("/api/admin/workbooks")
-                    .then().log().all()
-                    .extract();
-        }
-    }
-
-    public void 여러개_카드_생성_요청(List<AdminCardRequest> requests) {
-        for (AdminCardRequest request : requests) {
-            RestAssured.given().log().all()
-                    .contentType(MediaType.APPLICATION_JSON_VALUE)
-                    .auth().oauth2(로그인되어_있음().getAccessToken())
-                    .body(request)
-                    .when().post("/api/admin/cards")
-                    .then().log().all()
-                    .extract();
-        }
-    }
 }
